Remove `include/grpc/impl/codegen/slice.h`
