+++ BUILD
@@ -215,6 +215,7 @@ GRPC_PUBLIC_HDRS = [
     "include/grpc/impl/codegen/slice.h",
     "include/grpc/impl/compression_types.h",
     "include/grpc/impl/connectivity_state.h",
+    "include/grpc/impl/slice_type.h",
 ]
 
 GRPC_PUBLIC_EVENT_ENGINE_HDRS = [
