+++ include/grpc/event_engine/slice_buffer.h
@@ -26,7 +26,6 @@
 #include "absl/utility/utility.h"
 
 #include <grpc/event_engine/slice.h>
-#include <grpc/impl/codegen/slice.h>
 #include <grpc/slice.h>
 #include <grpc/slice_buffer.h>
 #include <grpc/support/log.h>
